+++ app-backend/db/src/main/resources/migrations/V8__Cascade_delete_task_action.sql
@@ -1,0 +1,5 @@
+-- Cascade the deletion of task actions when tasks are deleted
+
+ALTER TABLE public.task_actions
+DROP CONSTRAINT "task_actions_task_id_fkey",
+ADD CONSTRAINT "task_actions_task_id_fkey" FOREIGN KEY (task_id) REFERENCES tasks(id) ON DELETE CASCADE;